+++ assets/characters/angler-front-moving.png
