+++ assets/characters/angler-left-moving.png
