+++ assets/characters/angler-left-still.png
