+++ assets/characters/angler-right-moving.png
