+++ assets/sprite_sheet.png
