+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod camera;
+pub mod player;
+pub mod collision;
+pub mod resources;
